refined pointers

+++ src/StackFrame.hpp
@@ -9,7 +9,7 @@ typedef std::map<std::string, Variable> Scope;
 // memory locations are sort of an alternative to pointers, where a memory location contains the scopes location and the variables location within that scope
 // they can be encoded so that it contains both the index of the scope of the current stack frame and the index of the variable inside that scope
 // this makes it easily possible to get a variable from within a stack frame
-typedef uint32_t MemoryLocation; // memory location is defined so that its easier to change the type later on 
+typedef uint64_t MemoryLocation; // memory location is defined so that its easier to change the type later on 
 
 class StackFrame
 {
